Change version folders from `1.x` to `v1.x` and fix some urls to allow daux.io generation.
Daux.io doesn't like folders starting with a number (it treats those as the sortorder), so the folders had to be renamed causing the big diff. This commit also fixes some links which daux.io flagged as incorrect.

For SimpleCart specifically not all files are committed yet (as they're missing content), so trying to build with `daux` may not immediately work. These files will be committed when they have some content.