Set up multilingual "support"
This primarily accomplishes that the language is not shown in the breadcrumbs (yet links remain in-tact!), and of course in the future this will allow us to add additional languages that behave properly. We will need to add some way to switch languages at that point, but that's a later concern.